pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp-use>=1.2.8",
    "mcp[cli]>=1.7.1",
    "python-dotenv>=1.1.0",
]
